Rename package back to opensearch
